refactor: 체스 말 움직임 전략 중복 코드 제거
